ai-engineering-toolkit/src/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
pymongo==4.6.0
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.6
redis==5.0.1

# LLM & AI
openai==1.3.5
anthropic==0.3.11
cohere==4.32
instructor==0.4.5
langchain==0.0.350
llama-index==0.9.11
dspy==2.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Template Engine
Jinja2==3.1.2

# Vector Databases
pinecone-client==2.2.4
chromadb==0.4.17

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
prometheus-client==0.19.0
logging==0.4.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1